+++ lib/FRONT_END/authentications/auth_classes/verifying_email.dart
