+++ lib/FRONT_END/screens/option_components/support_resources.dart
