+++ lib/FRONT_END/screens/terms_page.dart
