--- lib/authentications/member_auth/member_login_page.dart
@@ -1,180 +1,0 @@
-
-import 'package:bethel_app_final/authentications/auth_classes/class_page.dart';
-import 'package:bethel_app_final/authentications/forgot_password.dart';
-import 'package:bethel_app_final/colors/color.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-
-class MemberLoginPage extends StatefulWidget {
-  final void Function()? onTap;
-
-  const MemberLoginPage({
-    Key? key,
-    this.onTap,
-  }) : super(key: key);
-
-  @override
-  State<MemberLoginPage> createState() => _MemberLoginPageState();
-}
-
-class _MemberLoginPageState extends State<MemberLoginPage> {
-  final emailController = TextEditingController();
-  final passwordController = TextEditingController();
-
-  void signUserIn() async {
-    try {
-      String email = emailController.text.trim();
-      String password = passwordController.text.trim();
-
-      if (email.isEmpty && password.isEmpty) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('Please enter both email and password.'),
-            duration: Duration(seconds: 3),
-          ),
-        );
-        return;
-      }
-      if (email.isEmpty) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('Please enter your email.'),
-            duration: Duration(seconds: 3),
-          ),
-        );
-        return;
-      }
-      if (password.isEmpty) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('Please enter your password.'),
-            duration: Duration(seconds: 3),
-          ),
-        );
-        return;
-      }
-
-      await FirebaseAuth.instance.signInWithEmailAndPassword(
-        email: email,
-        password: password,
-      );
-    } on FirebaseAuthException catch (e) {
-      print('FirebaseAuthException occurred: ${e.message}');
-      if (e.code == 'wrong-password') {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content:
-                Text('Incorrect password. Please double-check and try again.'),
-            duration: Duration(seconds: 3),
-          ),
-        );
-      } else if (e.code == 'user-not-found') {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('Email address not found. Please create an account.'),
-            duration: Duration(seconds: 3),
-          ),
-        );
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(title: const Text(''), automaticallyImplyLeading: true),
-      body: SafeArea(
-        child: Center(
-          child: SingleChildScrollView(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const SizedBox(height: 2),
-                Image.asset(
-                  'assets/images/churchmepicture.png',
-                  width: 400,
-                  height: 250,
-                ),
-                const SizedBox(height: 15),
-                const Padding(
-                  padding: EdgeInsets.only(
-                      right: 170), // Adjust the left padding as needed
-                  child: Text(
-                    'Member Login',
-                    style: TextStyle(
-                      color: Colors.black87,
-                      fontSize: 26,
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 20),
-                MyTextField(
-                  controller: emailController,
-                  hintText: 'Email',
-                  obscureText: false,
-                ),
-                const SizedBox(height: 10),
-                MyTextField(
-                  controller: passwordController,
-                  hintText: 'Password',
-                  obscureText: true,
-                ),
-                const SizedBox(height: 10),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 25.0),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.end,
-                    children: [
-                      GestureDetector(
-                        onTap: () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => const ForgotPassword(),
-                            ),
-                          );
-                        },
-                        child: const Text(
-                          'Forgot Password?',
-                          style: TextStyle(
-                            color: appGreen,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 26),
-                MyButton(
-                  onTap: signUserIn,
-                ),
-                const SizedBox(height: 10),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(
-                      "Don't have an account?",
-                      style: TextStyle(color: Colors.grey[700]),
-                    ),
-                    const SizedBox(width: 4),
-                    GestureDetector(
-                      onTap: widget.onTap,
-                      child: const Text(
-                        'Register now',
-                        style: TextStyle(
-                          color: appGreen,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    )
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
